+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# Unreleased
+
+Added
+ * AZ-832: Adding Zones and Private IPs range management
+
 # v6.0.0 - 2022-05-20
 
 Breaking